V1 Endpoints implemented and tested

+++ src/main/java/com/bookit/backend/payload/ShowRequest.java
@@ -31,7 +31,7 @@ public class ShowRequest {
 
     @NotNull
     @Positive
-    private Double basePrice;
+    private BigDecimal basePrice;
 
     private ShowStatus status;
